Improvements
Changed admin id from int to long (also in the example) since telegram user ids are not integers
Instead of returning null when we can't find a user in RemoveBotAdmin and AddBotAdmin let's return "Unknown user or user is not cached!" (because we are checking the database for the target user) or "You can't add/remove yourself!"
Changed "GetTarget" so it returns null instead of returning the source user (which made the null check pointless)
Made file logging optional (added an option in DefaultBotOptions and the interface)
Made loading default modules optional
Made loading addon modules optional
Fixed a few typos

+++ TelegramBotFramework.Core/SQLiteDb/TelegramBotSetting.cs
@@ -19,7 +19,7 @@ namespace TelegramBotFramework.Core.SQLiteDb
         /// <summary>
         /// The ID of the Telegram user who will be the main admin for the bot (typically, the person running the code)
         /// </summary>
-        public int TelegramDefaultAdminUserId { get; set; }
+        public long TelegramDefaultAdminUserId { get; set; }
         /// <summary>
         /// Your Telegram Bot API Token
         /// </summary>
